# shared/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "OverUnder API", "version": "0.1.0" },
  "paths": {
...
    "/api/v1/aa/userop": {
      "post": {
        "tags": ["aa"],
        "summary": "Sponsor Userop",
        "operationId": "sponsor_userop_api_v1_aa_userop_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [{ "$ref": "#/components/schemas/UserOpRequest" }, { "type": "null" }],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Successful Response", "content": { "application/json": { "schema": {} } } },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
    "/api/v1/aa/cdp-send": {
      "post": {
        "tags": ["aa"],
        "summary": "Cdp Send",
        "operationId": "cdp_send_api_v1_aa_cdp_send_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Authorization" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CdpSendRequest" } } }
        },
        "responses": {
          "200": { "description": "Successful Response", "content": { "application/json": { "schema": {} } } },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CdpCall": {
        "properties": {
          "to": { "type": "string", "title": "To" },
          "data": { "type": "string", "title": "Data" },
          "value": { "anyOf": [{ "type": "integer" }, { "type": "string" }], "title": "Value", "default": 0 }
        },
        "type": "object",
        "required": ["to", "data"],
        "title": "CdpCall"
      },
...
      "CdpSendRequest": {
        "properties": {
          "calls": { "items": { "$ref": "#/components/schemas/CdpCall" }, "type": "array", "title": "Calls" },
          "address": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Address" }
        },
        "type": "object",
        "required": ["calls"],
        "title": "CdpSendRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": { "items": { "$ref": "#/components/schemas/ValidationError" }, "type": "array", "title": "Detail" }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UserOpRequest": {
        "properties": {
          "sender": { "type": "string", "title": "Sender", "default": "" },
          "nonce": { "type": "integer", "title": "Nonce", "default": 0 },
          "initCode": { "type": "string", "title": "Initcode", "default": "0x" },
          "callData": { "type": "string", "title": "Calldata", "default": "0x" },
          "accountGasLimits": { "type": "string", "title": "Accountgaslimits", "default": "0x" },
          "preVerificationGas": { "type": "integer", "title": "Preverificationgas", "default": 0 },
          "gasFees": { "type": "string", "title": "Gasfees", "default": "0x" },
          "paymasterAndData": { "type": "string", "title": "Paymasteranddata", "default": "0x" },
          "signature": { "type": "string", "title": "Signature", "default": "0x" }
        },
        "type": "object",
        "title": "UserOpRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" },
          "input": { "title": "Input" },
          "ctx": { "type": "object", "title": "Context" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    }
  }
}